@@ -15,38 +15,41 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
+        "php": "^7.3|^8.0",
+        "ext-json": "*",
         "ext-mbstring": "*",
         "ext-openssl": "*",
-        "doctrine/inflector": "^1.1",
-        "dragonmantank/cron-expression": "^2.0",
-        "erusev/parsedown": "^1.7",
-        "laravel/nexmo-notification-channel": "^1.0",
-        "laravel/slack-notification-channel": "^1.0",
-        "league/flysystem": "^1.0.8",
-        "monolog/monolog": "^1.12",
-        "nesbot/carbon": "^1.26.3",
-        "opis/closure": "^3.1",
+        "doctrine/inflector": "^1.4|^2.0",
+        "dragonmantank/cron-expression": "^3.0.2",
+        "egulias/email-validator": "^2.1.10",
+        "league/commonmark": "^1.3|^2.0",
+        "league/flysystem": "^1.1",
+        "monolog/monolog": "^2.0",
+        "nesbot/carbon": "^2.31",
+        "opis/closure": "^3.6",
         "psr/container": "^1.0",
         "psr/simple-cache": "^1.0",
-        "ramsey/uuid": "^3.7",
+        "ramsey/uuid": "^4.0",
         "swiftmailer/swiftmailer": "^6.0",
-        "symfony/console": "^4.1",
-        "symfony/debug": "^4.1",
-        "symfony/finder": "^4.1",
-        "symfony/http-foundation": "^4.1",
-        "symfony/http-kernel": "^4.1",
-        "symfony/process": "^4.1",
-        "symfony/routing": "^4.1",
-        "symfony/var-dumper": "^4.1",
-        "tijsverkoyen/css-to-inline-styles": "^2.2.1",
-        "vlucas/phpdotenv": "^2.2"
+        "symfony/console": "^5.1.4",
+        "symfony/error-handler": "^5.1.4",
+        "symfony/finder": "^5.1.4",
+        "symfony/http-foundation": "^5.1.4",
+        "symfony/http-kernel": "^5.1.4",
+        "symfony/mime": "^5.1.4",
+        "symfony/process": "^5.1.4",
+        "symfony/routing": "^5.1.4",
+        "symfony/var-dumper": "^5.1.4",
+        "tijsverkoyen/css-to-inline-styles": "^2.2.2",
+        "vlucas/phpdotenv": "^5.2",
+        "voku/portable-ascii": "^1.4.8"
     },
     "replace": {
         "illuminate/auth": "self.version",
         "illuminate/broadcasting": "self.version",
         "illuminate/bus": "self.version",
         "illuminate/cache": "self.version",
+        "illuminate/collections": "self.version",
         "illuminate/config": "self.version",
         "illuminate/console": "self.version",
         "illuminate/container": "self.version",
