@@ -14,11 +14,13 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
-        "illuminate/contracts": "5.7.*",
-        "illuminate/http": "5.7.*",
-        "illuminate/queue": "5.7.*",
-        "illuminate/support": "5.7.*"
+        "php": "^7.3|^8.0",
+        "illuminate/collections": "^8.0",
+        "illuminate/contracts": "^8.0",
+        "illuminate/http": "^8.0",
+        "illuminate/macroable": "^8.0",
+        "illuminate/queue": "^8.0",
+        "illuminate/support": "^8.0"
     },
     "autoload": {
         "psr-4": {
@@ -27,13 +29,13 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.7-dev"
+            "dev-master": "8.x-dev"
         }
     },
     "suggest": {
-        "illuminate/console": "Required to use the auth:clear-resets command (5.7.*).",
-        "illuminate/queue": "Required to fire login / logout events (5.7.*).",
-        "illuminate/session": "Required to use the session based guard (5.7.*)."
+        "illuminate/console": "Required to use the auth:clear-resets command (^8.0).",
+        "illuminate/queue": "Required to fire login / logout events (^8.0).",
+        "illuminate/session": "Required to use the session based guard (^8.0)."
     },
     "config": {
         "sort-packages": true
