@@ -14,12 +14,14 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
+        "php": "^7.3|^8.0",
+        "ext-json": "*",
         "psr/log": "^1.0",
-        "illuminate/bus": "5.7.*",
-        "illuminate/contracts": "5.7.*",
-        "illuminate/queue": "5.7.*",
-        "illuminate/support": "5.7.*"
+        "illuminate/bus": "^8.0",
+        "illuminate/collections": "^8.0",
+        "illuminate/contracts": "^8.0",
+        "illuminate/queue": "^8.0",
+        "illuminate/support": "^8.0"
     },
     "autoload": {
         "psr-4": {
@@ -28,11 +30,11 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.7-dev"
+            "dev-master": "8.x-dev"
         }
     },
     "suggest": {
-        "pusher/pusher-php-server": "Required to use the Pusher broadcast driver (^3.0)."
+        "pusher/pusher-php-server": "Required to use the Pusher broadcast driver (^4.0|^5.0|^6.0)."
     },
     "config": {
         "sort-packages": true
