@@ -14,9 +14,11 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
-        "illuminate/contracts": "5.7.*",
-        "illuminate/support": "5.7.*"
+        "php": "^7.3|^8.0",
+        "illuminate/collections": "^8.0",
+        "illuminate/contracts": "^8.0",
+        "illuminate/macroable": "^8.0",
+        "illuminate/support": "^8.0"
     },
     "autoload": {
         "psr-4": {
@@ -25,13 +27,15 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "5.7-dev"
+            "dev-master": "8.x-dev"
         }
     },
     "suggest": {
-        "illuminate/database": "Required to use the database cache driver (5.7.*).",
-        "illuminate/filesystem": "Required to use the file cache driver (5.7.*).",
-        "illuminate/redis": "Required to use the redis cache driver (5.7.*)."
+        "ext-memcached": "Required to use the memcache cache driver.",
+        "illuminate/database": "Required to use the database cache driver (^8.0).",
+        "illuminate/filesystem": "Required to use the file cache driver (^8.0).",
+        "illuminate/redis": "Required to use the redis cache driver (^8.0).",
+        "symfony/cache": "Required to PSR-6 cache bridge (^5.1.4)."
     },
     "config": {
         "sort-packages": true
