@@ -1,5 +1,9 @@
 <?php
 
+use App\Http\Controllers\ExcelUploadController;
+use App\Http\Controllers\MapController;
+use Illuminate\Support\Facades\Route;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -11,19 +15,17 @@
 |
 */
 
-Route::post('excel_Submit', 'ExcelUploadController@excel')->name('excel.upload');
-Route::get('/execelUpload', 'ExcelUploadController@index')->name('excel.index');
+Route::post('excel_Submit', [ExcelUploadController::class, 'excel'])->name('excel.upload');
+Route::get('/execelUpload', [ExcelUploadController::class, 'index'])->name('excel.index');
 
-Route::post('/execelUpload/update/', 'ExcelUploadController@updateSubmit')->name('excel.update');
-Route::post('/execelUpload/delete', 'ExcelUploadController@delete')->name('excel.delete');
-Route::get('/execelUpload/update/{id}', 'ExcelUploadController@update');
+Route::post('/execelUpload/update/', [ExcelUploadController::class, 'updateSubmit'])->name('excel.update');
+Route::post('/execelUpload/delete', [ExcelUploadController::class, 'delete'])->name('excel.delete');
+Route::get('/execelUpload/update/{id}', [ExcelUploadController::class, 'update']);
+Route::post('/execelUpload/allDelete', [ExcelUploadController::class, 'allDelete'])->name('excel.all_delete');
 
-Route::post('/execelUpload/allDelete', 'ExcelUploadController@allDelete')->name('excel.all_delete');
-Route::post('/execelUpload/delete', 'ExcelUploadController@delete')->name('excel.delete');
-
-Route::get('/', 'MapController@index')->name('map.index');
-Route::post('/search', 'MapController@search')->name('search');
-Route::get('/mobile_search', 'MapController@mobileSearch')->name('mobile_search');
-Route::post('/mobile_search/result', 'MapController@mobileSearchResult')->name('mobile_result');
-Route::get('/map_list', 'MapController@list')->name('map.list');
-Route::get('/map/detail/{id}', 'MapController@detail')->name('map.detail');
+Route::get('/', [MapController::class, 'index'])->name('map.index');
+Route::post('/search', [MapController::class, 'search'])->name('search');
+Route::get('/mobile_search', [MapController::class, 'mobileSearch'])->name('mobile_search');
+Route::post('/mobile_search/result', [MapController::class, 'mobileSearchResult'])->name('mobile_result');
+Route::get('/map_list', [MapController::class, 'list'])->name('map.list');
+Route::get('/map/detail/{id}', [MapController::class, 'detail'])->name('map.detail');
